pyproject.toml
[project]
name = "westmarch-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "discord-py>=2.6.4",
    "dotenv>=0.9.9",
    "google>=3.0.0",
    "gspread>=6.2.1",
    "oauth2client>=4.1.3",
    "pytest>=8.4.2",
    "pytest-asyncio>=1.2.0",
]
